Add project configuration
